api/requirements.txt
django-rq==2.4.0
djangorestframework==3.12.2
whitenoise==5.3.0
requests==2.25.1
lxml==4.6.3
dateparser==1.0.0
gunicorn==20.0.4
drf-yasg[validation]

# ..................................................................
# For background work (async work)
celery==5.1.2
redis==3.5.3